Moved vignette from inst/doc/ to vignettes/ directory.
Small touch-ups in vignette: cache <- FALSE again and set.seed(1090)
added for chess22 example. The latter is also added in the replication
materials for JSS.

Test output for the vignette and examples is added.

CITATION added (with EEEcon working paper).

DESCRIPTION/NEWS updated accordings and ready for new CRAN release.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/partykit@291 255f4249-2263-4417-b1ea-8498779e27a7